Moved scipy loaders to a new file, now loading with register_class()
